backend_django/requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
mysqlclient==2.2.4
python-dotenv==1.2.1
sqlparse==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.153.0
